Safari-611.1.10/LayoutTests/http/wpt/webauthn/public-key-credential-get-failure-nfc.https-expected.txt

CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.

PASS PublicKeyCredential's [[get]] with timeout in a mock nfc authenticator.
